+++ config/fcos/v1_6_exp/validate.go
@@ -22,6 +22,23 @@ import (
 	"github.com/coreos/vcontext/report"
 )
 
+const rootDevice = "/dev/disk/by-id/coreos-boot-disk"
+
+// We can't define a Validate function directly on Disk because that's defined in base,
+// so we use a Validate function on the top-level Config instead.
+func (conf Config) Validate(c path.ContextPath) (r report.Report) {
+	for i, disk := range conf.Storage.Disks {
+		if disk.Device != rootDevice && !util.IsTrue(disk.WipeTable) {
+			for p, partition := range disk.Partitions {
+				if partition.Number == 0 && partition.Label != nil {
+					r.AddOnWarn(c.Append("storage", "disks", i, "partitions", p, "number"), common.ErrReuseByLabel)
+				}
+			}
+		}
+	}
+	return
+}
+
 func (d BootDevice) Validate(c path.ContextPath) (r report.Report) {
 	if d.Layout != nil {
 		switch *d.Layout {